started to add typing and moved stuff to quemb.shared
